debugging: restored delete handler and service
